[refact] date관련 함수, inputbar 초기화 함수 개별 파일로 분할
