+++ App/src/js/date/getDate.js
@@ -1,0 +1,7 @@
+export const getDate = () => {
+  const year = document.getElementById("year");
+  const month = document.getElementById("month");
+  const month_en = document.getElementById("month_en");
+
+  return [year, month, month_en];
+};
